Rework acceptance tests to follow codebase pattern and merge it into a single file
